Move publisher node
